Complete REST API for the cooked schema: auth, users, ingredients,
recipes (moods/ingredients/instructions), pantry (FIFO deduction),
shopping list (upsert), cook history (kcal snapshot), saved recipes,
and vocabulary endpoints. JWT userId updated to BIGINT. 16 tests passing.

+++ src/main/java/com/humanworkstream/cooked/security/JwtUtil.java
@@ -24,7 +24,7 @@ public class JwtUtil {
         this.expirationMs = expirationMs;
     }
 
-    public String generate(String email, int userId) {
+    public String generate(String email, long userId) {
         return Jwts.builder()
                 .subject(email)
                 .claim("userId", userId)
